server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
pyngrok>=7.0.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0
cryptography>=41.0.0
Jinja2>=3.1.0
python-multipart>=0.0.6
aiofiles>=23.0.0
